pyproject.toml
[project]
name = "bdt-final-assignment"
version = "0.1.0"
description = "Final Assignment Big Data Tools and Techniques"
authors = [
    {name = "challa82@students.rowan.edu"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "cassandra-driver (>=3.29.2,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
